+++ CHANGELOG.md
@@ -6,10 +6,12 @@ Here is the list of the changes and features provided by the **Rainbow-Node-SDK*
 
 All notable changes to Rainbow-Node-SDK will be documented in this file.
 
-## [1.61.0] - 2018-09-xx
+## [1.61.0] - 2018-09-18
 -   Update the logs to remove all people data.
 -   CRRAINB-7686 : Fix code for the 'read' receipt sent automatically to the sender when the message is received.
--   Add the ability to start services one by one. To avoid the calls to unnecessary pay API on server. 
+-   Add the ability to start services one by one. To avoid the calls to unnecessary pay API on server.
+-   Add event `rainbow_onconferenced` fired when a conference event is received.
+    
 
 ## [1.60.0] - 2018-08-28
 -   Add method getStatusForUser in Bubble class to get the status of a user in the bubble